// src/interfaces/Lump.ts
export enum LumpName {
  FF_START = 'FF_START',
  F_START = 'F_START',
  F1_START = 'F1_START',
  F2_START = 'F2_START',
  F3_START = 'F3_START',
  SS_START = 'SS_START',
  S_START = 'S_START',
  P_START = 'P_START',
  P1_START = 'P1_START',
  P2_START = 'P2_START',
  P3_START = 'P3_START',
  FF_END = 'FF_END',
  F_END = 'F_END',
  F1_END = 'F1_END',
  F2_END = 'F2_END',
  F3_END = 'F3_END',
  SS_END = 'SS_END',
  S_END = 'S_END',
  P_END = 'P_END',
  P1_END = 'P1_END',
  P2_END = 'P2_END',
  P3_END = 'P3_END',
  BEHAVIOR = 'BEHAVIOR',
  PLAYPAL = 'PLAYPAL',
  COLORMAP = 'COLORMAP',
  ENDDOOM = 'ENDDOOM',
  PNAMES = 'PNAMES',
  TEXTURES = 'TEXTURES',
  TEXTURE1 = 'TEXTURE1',
  TEXTURE2 = 'TEXTURE2',
  GENMIDI = 'GENMIDI',
  DMXGUS = 'DMXGUS',
  DEMO1 = 'DEMO1',
  DEMO2 = 'DEMO2',
  DEMO3 = 'DEMO3',
  BLOCKMAP = 'BLOCKMAP',
  VERTEXES = 'VERTEXES',
  SECTORS = 'SECTORS',
  SIDEDEFS = 'SIDEDEFS',
  LINEDEFS = 'LINEDEFS',
  SSECTORS = 'SSECTORS',
  NODES = 'NODES',
  SEGS = 'SEGS',
  REJECT = 'REJECT',
  THINGS = 'THINGS',
}

export interface Lump {
  name: LumpName;
  data: ArrayBuffer;
}